requirements.txt
numpy>=1.21.0
pandas>=1.3.0
nltk>=3.6.0
spacy>=3.0.0
transformers>=4.10.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
requests>=2.26.0
Pystemmer>=3.0.0
spacy>=3.8.14